+++ src/o365/spo/commands/hubsite/AssociatedSite.ts
@@ -1,0 +1,6 @@
+export interface AssociatedSite {
+  HubSiteId?: string;
+  SiteId?: string;
+  SiteUrl: string;
+  Title: string;
+}