+++ src/o365/spo/commands/storageentity/storageentity-set.spec.ts
@@ -505,7 +505,7 @@ describe(commands.STORAGEENTITY_SET, () => {
   });
 
   it('rejects invalid SharePoint Online URL', () => {
-    const url = 'https://contoso.com';
+    const url = 'http://contoso';
     const actual = (command.validate() as CommandValidate)({ options: { appCatalogUrl: url } });
     assert.equal(actual, `${url} is not a valid SharePoint Online site URL`);
   });